Slightly restructure shared project
