+++ glyph-shared/src/main/kotlin/pubsub/redis/Redlock.kt
@@ -27,6 +27,7 @@ package org.yttr.glyph.shared.redis
 import io.lettuce.core.ScriptOutputType
 import io.lettuce.core.SetArgs
 import kotlinx.coroutines.future.await
+import org.yttr.glyph.shared.pubsub.redis.RedisAsync
 
 /**
  * Lock a Redlock locked value for a fixed duration.
